Added lots of misc. stuff.
Completed more of the Alf class, completed Skill stuff, added utilities,
completed more of Properties, completed a few events.

+++ ALFCore/src/com/alf/util/Util.java
@@ -110,7 +110,7 @@ public class Util {
 		 Integer newVal = toInt(val);
 		 if (newVal == null)
 			 throw new IllegalArgumentException(name + " must be a numeral!");
-		 return newVal.intValue();
+		 return newVal;
 	 }
 
 	 /** Convert parameter to a Double. Handles non-numeral case. */
@@ -118,7 +118,7 @@ public class Util {
 		 Double newVal = toDouble(val);
 		 if (newVal == null)
 			 throw new IllegalArgumentException(name + " must be a numeral!");
-		 return newVal.doubleValue();
+		 return newVal;
 	 }
 	 
 	 static {
